Moving all the SQL to common/SQL folder
