+++ common/SQL/CDT_COVID/Hospitalizations.sql
