+++ common/SQL/CDT_COVID/Oldstats.sql
